+++ templates/index.html
@@ -1,0 +1,66 @@
+{% extends "base_template.html" %}
+
+{% block content %}
+</head>
+<body>
+
+<div>
+
+<!--    <button id="saveGraphButton">Save Graph</button>-->
+<!--    <br>-->
+    <button id="triggerAgentsButton">Trigger Agents</button>
+    <br>
+    Central Gravity: <input id="centralGravity" type="range" min="0" max="1" step="0.01"><br>
+    Spring Length: <input id="springLength" type="range" min="1" max="1000"><br>
+    Spring Constant: <input id="springConstant" type="range" min="0" max="1" step="0.01"><br>
+    Node Distance: <input id="nodeDistance" type="range" min="1" max="1000"><br>
+    Damping: <input id="damping" type="range" min="0" max="1" step="0.01"><br>
+    Solver: <select id="physicsSolver">
+    <option value="hierarchicalRepulsion">Hierarchical Repulsion</option>
+    <option value="barnesHut">Barnes-Hut</option>
+    <option value="repulsion">Repulsion</option>
+    <option value="forceAtlas2Based">Force Atlas 2 Based</option>
+</select><br>
+    DWH: <select id="graphDropdown">
+    <option value="">Select a Graph</option>
+    <!-- Options will be populated dynamically from BQ -->
+</select><br>
+    Graph-Algo-Label: <select id="messagePassingDropdown">
+    <option value="">messagePassingDropdown</option>
+    <!-- Options will be populated dynamically from BQ -->
+</select>
+</div>
+
+
+<div id="graph-container">
+    <div id="graph"></div>
+    <div id="graph-controls">
+    </div>
+    <div id="editPanel">
+        <h3>Edit Node</h3>
+        <textarea id="nodeTextLabel" rows="4" style="width:100%"></textarea>
+        <br>
+        <button id="applyNodeChangesButton">Apply</button>
+        <button id="cancelNodeChangesButton">Cancel</button>
+    </div>
+
+    <div id="jsonDataPanel">
+        <h3>JSON Data</h3>
+        <textarea id="jsonData" rows="10" style="width:100%"></textarea>
+        <br>
+        <button id="updateFromJsonButton">Update Graph from JSON</button>
+        <button id="resetJsonButton">Reset JSON</button>
+    </div>
+
+    <script src="{{ url_for('static', filename='js/GraphUI.js') }}"></script>
+    <script>
+        document.addEventListener('DOMContentLoaded', (event) => {
+            const graphUI = new GraphUI();
+            graphUI.init();
+        });
+    </script>
+</div>
+
+</body>
+</html>
+{% endblock %}
